refactor: move `get_raw_conversation()` into `Client` module [WPB-15530]
